moves all api to companies.ex from companies/companies.ex
