--- lib/companies_web/templates/company/live_browse_companies.html.leex
@@ -1,22 +1,0 @@
-<section class="section">
-  <div class="container">
-
-    <h1 class="title is-3 fancy"><%= gettext("Browse") %></h1>
-
-    <div class="tabs is-centered">
-      <form action="">
-      </form>
-    </div>
-
-    <div class="content">
-      <div class="columns is-mobile is-multiline">
-        <%= for company <- @companies do %>
-          <div class="column is-full-mobile is-half-tablet is-one-third-desktop is-one-quarter-widescreen">
-            <%= render "livecard.html", company: company %>
-          </div>
-        <% end %>
-      </div> <!-- /columns -->
-    </div> <!-- /content -->
-
-  </div> <!-- /container -->
-</section>
